requirements.txt
flask==2.0.1
PyJWT==2.3.0
werkzeug==2.0.2
gunicorn==20.1.0
python-dotenv==0.19.2
cryptography==36.0.1